+++ src/main/java/edu/nechaev/project/repositories/CommentPostRepository.java
@@ -10,8 +10,8 @@ import org.springframework.stereotype.Repository;
 import java.sql.Timestamp;
 
 @Repository
-public interface CommentPostRepository extends CrudRepository<CommentPost, Integer> {
+public interface CommentPostRepository extends CrudRepository<CommentPost, Long> {
     @Modifying
     @Query("UPDATE comment SET text = :text, time = :time, is_changed = :isChanged WHERE id = :id")
-    int updateCommentById(int id, String text, Timestamp time, boolean isChanged);
+    long updateCommentById(long id, String text, Timestamp time, boolean isChanged);
 }
