+++ src/main/java/edu/nechaev/project/repositories/ProgramRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProgramRepository extends CrudRepository<Program, Integer> {
+public interface ProgramRepository extends CrudRepository<Program, Long> {
 }
