+++ src/main/java/edu/nechaev/project/repositories/UserRoleRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRoleRepository extends CrudRepository<UserRole, Integer> {
+public interface UserRoleRepository extends CrudRepository<UserRole, Long> {
     Iterable<UserRole> findByRole(String role);
 }
